+++ CHANGELOG.md
@@ -6,6 +6,12 @@ ByteriX, 2017. All right reserved.
 
 # Versions
 
+## 2.4.3 (31.07.2017)
+##### Improvements
+* fixed BxInputEmptyValueChecker for BxInputFormattedTextRow
+* added BxInputBlockChecker for universal checking
+* added errorColor to BxInputController for global using BxInputStandartErrorRowDecorator
+
 ## 2.4.2 (26.07.2017)
 ##### Bug fix
 * change select/deselect row logic with bug fixes